SendWithStartTLS: Allow to set TLS nil
Setting TLS nil will skip STARTTLS and with that do better than the stdlib when smtp.SendEmail is used. There, the STARTLS announced by the server will always result in starting TLS. In our case at hand, the server does announce STARTTLS but has no properly signed certificates. So in order to send an email with SKIP tls we need this patch.